Mqtt Actor fix and add support for tasmota sonoff
There was an error when the performance was changed. I've replaced round () with int (). Now the percentage of the performance is displayed on the icon of the actor.
Added support for Sonoff flashed with Tasmota.
Payload for ESPEasy added. But I couldn't test it. The same with the original payload used by alexander. I can't test it.
I tested whether changing configurations lead to errors. But I cannot test the functionality. Sonoff is of course tested